matrixMultiplication is a double now

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Skystone/MathFunctions.java
@@ -81,11 +81,11 @@ public class MathFunctions {
         return java.lang.Math.sqrt(sum);
     }
 
-    public static int[][] matrixMultiplication(int[][] a, int[][] b) {
+    public static double[][] matrixMultiplication(double[][] a, double[][] b) {
         int rowsInA = a.length;
         int columnsInA = a[0].length; // same as rows in B
         int columnsInB = b[0].length;
-        int[][] c = new int[rowsInA][columnsInB];
+        double[][] c = new int[rowsInA][columnsInB];
         for (int i = 0; i < rowsInA; i++) {
             for (int j = 0; j < columnsInB; j++) {
                 for (int k = 0; k < columnsInA; k++) {
